@@ -1,6 +1,6 @@
 package ht.treechop.api;
 
-import net.minecraft.block.BlockState;
+import net.minecraft.world.level.block.state.BlockState;
 
 public interface IChoppableBlock {
 
