@@ -1,7 +1,7 @@
 package ht.treechop.api;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.state.BlockState;
 
 public interface IChoppingItem {
 
