@@ -1,9 +1,11 @@
 package ht.treechop.client.gui.element;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
 import ht.treechop.client.gui.util.Sprite;
 import ht.treechop.client.gui.widget.SpriteButtonWidget;
-import net.minecraft.client.gui.IGuiEventListener;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
+import net.minecraft.network.chat.TextComponent;
 
 import java.util.Collections;
 import java.util.List;
@@ -13,19 +15,20 @@ public class ButtonGui extends NestedGui {
     private final SpriteButtonWidget widget;
 
     public ButtonGui(Sprite sprite, Sprite highlightedSprite, Runnable onPress) {
+        super(0, 0, 0, 0, TextComponent.EMPTY);
         this.widget = new SpriteButtonWidget(0, 0, sprite, highlightedSprite, onPress);
     }
 
     @Override
-    public List<? extends IGuiEventListener> getEventListeners() {
+    public List<? extends GuiEventListener> children() {
         return Collections.singletonList(widget);
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
         this.widget.x = getBox().getCenterX() - this.widget.getWidth() / 2;
-        this.widget.y = getBox().getCenterY() - this.widget.getHeightRealms() / 2;
-        this.widget.render(matrixStack, mouseX, mouseY, partialTicks);
+        this.widget.y = getBox().getCenterY() - this.widget.getHeight() / 2;
+        this.widget.render(poseStack, mouseX, mouseY, partialTicks);
     }
 
     @Override
@@ -35,7 +38,11 @@ public class ButtonGui extends NestedGui {
 
     @Override
     public int getMinimumHeight() {
-        return widget.getHeightRealms();
+        return widget.getHeight();
     }
 
+    @Override
+    public void updateNarration(NarrationElementOutput p_169152_) {
+        // TODO
+    }
 }
