@@ -1,98 +1,106 @@
-package ht.treechop.client.gui.element;
-
-import com.google.common.collect.Lists;
-import com.mojang.blaze3d.matrix.MatrixStack;
-import ht.treechop.client.gui.widget.StickyWidget;
-import net.minecraft.client.gui.IGuiEventListener;
-import net.minecraft.client.gui.widget.Widget;
-import net.minecraft.util.text.ITextComponent;
-
-import java.util.Collection;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-
-public class ExclusiveButtonsGui extends NestedGui {
-
-    private final List<Widget> widgets;
-
-    protected ExclusiveButtonsGui(Collection<Widget> widgets) {
-        this.widgets = Lists.newArrayList(widgets);
-    }
-
-    public void resize(int width) {
-        if (getMinimumWidth() < width) {
-            int targetWidth = width / widgets.size();
-            List<Widget> smallerWidgets = widgets.stream().filter(widget -> widget.getWidth() <= targetWidth).collect(Collectors.toList());
-            List<Widget> biggerWidgets = widgets.stream().filter(widget -> widget.getWidth() > targetWidth).collect(Collectors.toList());
-            int totalWidthForSmallers = width - biggerWidgets.stream().map(Widget::getWidth).reduce(Integer::sum).orElse(0);
-            int i = 0;
-            // Do this incrementally to account for rounding errors
-            for (Widget widget : smallerWidgets) {
-                double lower = (double) i / smallerWidgets.size();
-                double upper = (double) (i + 1) / smallerWidgets.size();
-                int widgetWidth = (int) (totalWidthForSmallers * (upper - lower));
-                widget.setWidth(widgetWidth);
-                ++i;
-            }
-        }
-    }
-
-    @SuppressWarnings("NullableProblems")
-    @Override
-    public List<? extends IGuiEventListener> getEventListeners() {
-        return widgets;
-    }
-
-    @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
-        int x = getBox().getLeft();
-        int y = getBox().getTop();
-        for (Widget widget : widgets) {
-            widget.x = x;
-            widget.y = y;
-            widget.render(matrixStack, mouseX, mouseY, partialTicks);
-            x += widget.getWidth();
-        }
-    }
-
-    @Override
-    public int getMinimumWidth() {
-        return widgets.stream().map(Widget::getWidth).reduce(Integer::sum).orElse(0);
-    }
-
-    @Override
-    public int getMinimumHeight() {
-        return widgets.stream().map(Widget::getHeightRealms).reduce(Integer::max).orElse(0);
-    }
-
-    public static class Builder {
-        private final List<Option> options = new LinkedList<>();
-
-        public Builder add(ITextComponent name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
-            options.add(new Option(name, onPress, stateSupplier));
-            return this;
-        }
-
-        public ExclusiveButtonsGui build() {
-            List<Widget> widgets = options.stream()
-                    .map(option -> new StickyWidget(0, 0, 0, 0, option.name, option.onPress, option.stateSupplier))
-                    .collect(Collectors.toList());
-            return new ExclusiveButtonsGui(widgets);
-        }
-
-        private static class Option {
-            private ITextComponent name;
-            private Runnable onPress;
-            private Supplier<StickyWidget.State> stateSupplier;
-
-            public Option(ITextComponent name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
-                this.name = name;
-                this.onPress = onPress;
-                this.stateSupplier = stateSupplier;
-            }
-        }
-    }
-
-}
+//package ht.treechop.client.gui.element;
+//
+//import com.google.common.collect.Lists;
+//import com.mojang.blaze3d.vertex.PoseStack;
+//import ht.treechop.client.gui.widget.StickyWidget;
+//import net.minecraft.client.gui.components.AbstractWidget;
+//import net.minecraft.client.gui.components.events.GuiEventListener;
+//import net.minecraft.client.gui.narration.NarrationElementOutput;
+//import net.minecraft.network.chat.Component;
+//import net.minecraft.network.chat.TextComponent;
+//
+//import java.util.Collection;
+//import java.util.LinkedList;
+//import java.util.List;
+//import java.util.function.Supplier;
+//import java.util.stream.Collectors;
+//
+//public class ExclusiveButtonsGui extends NestedGui {
+//
+//    private final List<AbstractWidget> widgets;
+//
+//    protected ExclusiveButtonsGui(Collection<AbstractWidget> widgets) {
+//        super(0, 0, 0, 0, TextComponent.EMPTY);
+//        this.widgets = Lists.newArrayList(widgets);
+//    }
+//
+//    public void resize(int width) {
+//        if (getMinimumWidth() < width) {
+//            int targetWidth = width / widgets.size();
+//            List<AbstractWidget> smallerWidgets = widgets.stream().filter(widget -> widget.getWidth() <= targetWidth).collect(Collectors.toList());
+//            List<AbstractWidget> biggerWidgets = widgets.stream().filter(widget -> widget.getWidth() > targetWidth).collect(Collectors.toList());
+//            int totalWidthForSmallers = width - biggerWidgets.stream().map(AbstractWidget::getWidth).reduce(Integer::sum).orElse(0);
+//            int i = 0;
+//            // Do this incrementally to account for rounding errors
+//            for (AbstractWidget widget : smallerWidgets) {
+//                double lower = (double) i / smallerWidgets.size();
+//                double upper = (double) (i + 1) / smallerWidgets.size();
+//                int widgetWidth = (int) (totalWidthForSmallers * (upper - lower));
+//                widget.setWidth(widgetWidth);
+//                ++i;
+//            }
+//        }
+//    }
+//
+//    @SuppressWarnings("NullableProblems")
+//    @Override
+//    public List<? extends GuiEventListener> children() {
+//        return widgets;
+//    }
+//
+//    @Override
+//    public void render(PoseStack poseStack, int mouseX, int mouseY, float partialTicks) {
+//        int x = getBox().getLeft();
+//        int y = getBox().getTop();
+//        for (AbstractWidget widget : widgets) {
+//            widget.x = x;
+//            widget.y = y;
+//            widget.render(poseStack, mouseX, mouseY, partialTicks);
+//            x += widget.getWidth();
+//        }
+//    }
+//
+//    @Override
+//    public int getMinimumWidth() {
+//        return widgets.stream().map(AbstractWidget::getWidth).reduce(Integer::sum).orElse(0);
+//    }
+//
+//    @Override
+//    public int getMinimumHeight() {
+//        return widgets.stream().map(AbstractWidget::getHeight).reduce(Integer::max).orElse(0);
+//    }
+//
+//    @Override
+//    public void updateNarration(NarrationElementOutput out) {
+//        // TODO
+//    }
+//
+//    public static class Builder {
+//        private final List<Option> options = new LinkedList<>();
+//
+//        public Builder add(Component name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
+//            options.add(new Option(name, onPress, stateSupplier));
+//            return this;
+//        }
+//
+//        public ExclusiveButtonsGui build() {
+//            List<AbstractWidget> widgets = options.stream()
+//                    .map(option -> new StickyWidget(0, 0, 0, 0, option.name, option.onPress, option.stateSupplier))
+//                    .collect(Collectors.toList());
+//            return new ExclusiveButtonsGui(widgets);
+//        }
+//
+//        private static class Option {
+//            private Component name;
+//            private Runnable onPress;
+//            private Supplier<StickyWidget.State> stateSupplier;
+//
+//            public Option(Component name, Runnable onPress, Supplier<StickyWidget.State> stateSupplier) {
+//                this.name = name;
+//                this.onPress = onPress;
+//                this.stateSupplier = stateSupplier;
+//            }
+//        }
+//    }
+//
+//}
