@@ -1,21 +1,24 @@
 package ht.treechop.client.gui.screen;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.platform.GlStateManager;
+import com.mojang.blaze3d.platform.Window;
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import ht.treechop.client.Client;
 import ht.treechop.client.gui.util.Sprite;
 import ht.treechop.common.config.ConfigHandler;
 import ht.treechop.common.settings.ChopSettings;
 import ht.treechop.common.util.ChopUtil;
-import net.minecraft.client.MainWindow;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.entity.player.ClientPlayerEntity;
-import net.minecraft.client.gui.AbstractGui;
-import net.minecraft.client.world.ClientWorld;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.client.gui.GuiComponent;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraft.world.phys.HitResult;
 
-public class ChopIndicator extends AbstractGui {
+public class ChopIndicator extends GuiComponent {
 
     private static final double IMAGE_SCALE = 1.0;
 
@@ -23,20 +26,21 @@ public class ChopIndicator extends AbstractGui {
     private static final ChopSettings lastChopSettings = new ChopSettings();
     private static boolean lastBlockWouldBeChopped = false;
 
-    public void render(MainWindow window, MatrixStack matrixStack, float partialTicks) {
+    public void render(Window window, PoseStack poseStack, float partialTicks) {
         Minecraft minecraft = Minecraft.getInstance();
-        RayTraceResult mouseOver = minecraft.objectMouseOver;
+        HitResult mouseOver = minecraft.hitResult;
 
         if (Client.isChoppingIndicatorEnabled() &&
-                minecraft.currentScreen == null && mouseOver != null &&
-                mouseOver.getType() == RayTraceResult.Type.BLOCK &&
-                mouseOver instanceof BlockRayTraceResult
+                minecraft.screen == null && mouseOver != null &&
+                mouseOver.getType() == HitResult.Type.BLOCK &&
+                mouseOver instanceof BlockHitResult
         ) {
-            BlockPos blockPos = ((BlockRayTraceResult) mouseOver).getPos();
+            BlockPos blockPos = ((BlockHitResult) mouseOver).getBlockPos();
             if (blockWouldBeChopped(blockPos)) {
-                int windowWidth = window.getScaledWidth();
-                int windowHeight = window.getScaledHeight();
-                minecraft.getTextureManager().bindTexture(Sprite.TEXTURE_PATH);
+                RenderSystem.blendFuncSeparate(GlStateManager.SourceFactor.ONE_MINUS_DST_COLOR, GlStateManager.DestFactor.ONE_MINUS_SRC_COLOR, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
+                int windowWidth = window.getGuiScaledWidth();
+                int windowHeight = window.getGuiScaledHeight();
+                RenderSystem.setShaderTexture(0, Sprite.TEXTURE_PATH);
 
                 int indicatorCenterX = windowWidth / 2 + ConfigHandler.CLIENT.indicatorXOffset.get();
                 int indicatorCenterY = windowHeight / 2 + ConfigHandler.CLIENT.indicatorYOffset.get();
@@ -45,7 +49,7 @@ public class ChopIndicator extends AbstractGui {
                 int imageWidth = (int) (sprite.width * IMAGE_SCALE);
                 int imageHeight = (int) (sprite.height * IMAGE_SCALE);
                 sprite.blit(
-                        matrixStack,
+                        poseStack,
                         indicatorCenterX - imageWidth / 2,
                         indicatorCenterY - imageHeight / 2,
                         imageWidth,
@@ -61,21 +65,21 @@ public class ChopIndicator extends AbstractGui {
 
     private boolean blockWouldBeChopped(BlockPos pos) {
         Minecraft minecraft = Minecraft.getInstance();
-        ClientPlayerEntity player = minecraft.player;
-        ClientWorld world = minecraft.world;
+        LocalPlayer player = minecraft.player;
+        ClientLevel level = minecraft.level;
 
-        if (world != null & minecraft.player != null &&
-                ChopUtil.canChopWithTool(player.getHeldItemMainhand()) &&
+        if (level != null & minecraft.player != null &&
+                ChopUtil.canChopWithTool(player.getMainHandItem()) &&
                 ChopUtil.playerWantsToChop(minecraft.player, Client.getChopSettings())
         ) {
             if ((pos.equals(lastBlockPos) || !Client.getChopSettings().equals(lastChopSettings))
             ) {
                 if (ChopUtil.playerWantsToFell(player, Client.getChopSettings())) {
                     lastBlockWouldBeChopped = ChopUtil.isPartOfATree(
-                            world, pos, Client.getChopSettings().getTreesMustHaveLeaves()
+                            level, pos, Client.getChopSettings().getTreesMustHaveLeaves()
                     );
                 } else {
-                    lastBlockWouldBeChopped = ChopUtil.isBlockALog(world, pos);
+                    lastBlockWouldBeChopped = ChopUtil.isBlockALog(level, pos);
                 }
             }
         } else {
