@@ -1,16 +1,17 @@
 package ht.treechop.client.model;
 
+import com.mojang.math.Vector3f;
 import ht.treechop.common.util.FaceShape;
 import ht.treechop.common.util.Vector3;
-import net.minecraft.client.renderer.model.BakedQuad;
-import net.minecraft.client.renderer.model.BlockFaceUV;
-import net.minecraft.client.renderer.model.BlockPartFace;
-import net.minecraft.client.renderer.model.FaceBakery;
+import net.minecraft.client.renderer.block.model.BakedQuad;
+import net.minecraft.client.renderer.block.model.BlockElementFace;
+import net.minecraft.client.renderer.block.model.BlockFaceUV;
+import net.minecraft.client.renderer.block.model.FaceBakery;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.util.Direction;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.vector.Vector3f;
-import net.minecraftforge.client.model.SimpleModelTransform;
+import net.minecraft.client.resources.model.ModelState;
+import net.minecraft.core.Direction;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.client.model.SimpleModelState;
 
 public class ModelUtil {
     public static BakedQuad makeQuad(
@@ -56,14 +57,14 @@ public class ModelUtil {
         switch (orientation) {
             case UP:
             case DOWN:
-                return new float[]{posFrom3f.getX(), posFrom3f.getZ(), posTo3f.getX(), posTo3f.getZ()};
+                return new float[]{posFrom3f.x(), posFrom3f.z(), posTo3f.x(), posTo3f.z()};
             case EAST:
             case WEST:
-                return new float[]{posFrom3f.getZ(), posFrom3f.getY(), posTo3f.getZ(), posTo3f.getY()};
+                return new float[]{posFrom3f.z(), posFrom3f.y(), posTo3f.z(), posTo3f.y()};
             case NORTH:
             case SOUTH:
             default:
-                return new float[]{posFrom3f.getX(), posFrom3f.getY(), posTo3f.getX(), posTo3f.getY()};
+                return new float[]{posFrom3f.x(), posFrom3f.y(), posTo3f.x(), posTo3f.y()};
         }
     }
 
@@ -80,10 +81,10 @@ public class ModelUtil {
         return new FaceBakery().bakeQuad(
                 posFrom.asVector3f(),
                 posTo.asVector3f(),
-                new BlockPartFace(culling, -1, textureRL.toString(), new BlockFaceUV(uvs, uvRotation)),
+                new BlockElementFace(culling, -1, textureRL.toString(), new BlockFaceUV(uvs, uvRotation)),
                 sprite,
                 orientation,
-                SimpleModelTransform.IDENTITY,
+                SimpleModelState.IDENTITY,
                 null,
                 true,
                 null
