@@ -3,19 +3,17 @@ package ht.treechop.common.block;
 import ht.treechop.api.IChoppableBlock;
 import ht.treechop.common.properties.BlockStateProperties;
 import ht.treechop.common.properties.ChoppedLogShape;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.state.EnumProperty;
-import net.minecraft.state.IntegerProperty;
-import net.minecraft.state.StateContainer;
-import net.minecraft.util.Direction;
-import net.minecraft.util.Direction.Axis;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorld;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.shapes.Shapes;
+import net.minecraft.world.phys.shapes.VoxelShape;
 
 import static ht.treechop.common.util.ChopUtil.isBlockALog;
 import static ht.treechop.common.util.ChopUtil.isBlockLeaves;
@@ -26,15 +24,17 @@ public class ChoppedLogBlock extends Block implements IChoppableBlock {
     protected static final EnumProperty<ChoppedLogShape> SHAPE = BlockStateProperties.CHOPPED_LOG_SHAPE;
 
     public ChoppedLogBlock(Properties properties) {
-        super(properties.variableOpacity());
-        this.setDefaultState(
-                this.stateContainer.getBaseState()
-                        .with(CHOPS, 1)
-                        .with(SHAPE, ChoppedLogShape.PILLAR_Y)
+        super(properties
+                .dynamicShape()
+                .isViewBlocking((BlockState blockState, BlockGetter level, BlockPos pos) -> false));
+        this.registerDefaultState(
+                this.getStateDefinition().any()
+                        .setValue(CHOPS, 1)
+                        .setValue(SHAPE, ChoppedLogShape.PILLAR_Y)
         );
     }
 
-    public static ChoppedLogShape getPlacementShape(IWorld world, BlockPos blockPos) {
+    public static ChoppedLogShape getPlacementShape(Level level, BlockPos blockPos) {
         final byte DOWN     = 1;
         final byte UP       = 1 << 1;
         final byte NORTH    = 1 << 2;
@@ -43,50 +43,44 @@ public class ChoppedLogBlock extends Block implements IChoppableBlock {
         final byte EAST     = 1 << 5;
 
         byte openSides = (byte) (
-                (isBlockOpen(world, blockPos.down()) ? DOWN : 0)
-                | (!isBlockALog(world, blockPos.up()) ? UP : 0)
-                | (!isBlockALog(world, blockPos.north()) ? NORTH : 0)
-                | (!isBlockALog(world, blockPos.south()) ? SOUTH : 0)
-                | (!isBlockALog(world, blockPos.west()) ? WEST : 0)
-                | (!isBlockALog(world, blockPos.east()) ? EAST : 0)
+                (isBlockOpen(level, blockPos.below()) ? DOWN : 0)
+                | (!isBlockALog(level, blockPos.above()) ? UP : 0)
+                | (!isBlockALog(level, blockPos.north()) ? NORTH : 0)
+                | (!isBlockALog(level, blockPos.south()) ? SOUTH : 0)
+                | (!isBlockALog(level, blockPos.west()) ? WEST : 0)
+                | (!isBlockALog(level, blockPos.east()) ? EAST : 0)
         );
 
         return ChoppedLogShape.forOpenSides(openSides);
     }
 
-    private static boolean isBlockOpen(IWorld world, BlockPos pos) {
-        return (world.isAirBlock(pos.down()) || isBlockLeaves(world, pos.down()));
+    private static boolean isBlockOpen(Level level, BlockPos pos) {
+        return (level.isEmptyBlock(pos.below()) || isBlockLeaves(level, pos.below()));
+    }
+
+    @Override
+    public boolean hasDynamicShape() {
+        return true;
     }
 
     @SuppressWarnings({"deprecation", "NullableProblems"})
     @Override
-    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
-        int chops = state.get(CHOPS);
-        AxisAlignedBB box = state.get(SHAPE).getBoundingBox(chops);
-        VoxelShape voxelShape = Block.makeCuboidShape(
-                box.getMin(Axis.X),
-                box.getMin(Axis.Y),
-                box.getMin(Axis.Z),
-                box.getMax(Axis.X),
-                box.getMax(Axis.Y),
-                box.getMax(Axis.Z)
+    public VoxelShape getOcclusionShape(BlockState state, BlockGetter level, BlockPos pos) {
+        int chops = state.getValue(CHOPS);
+        AABB box = state.getValue(SHAPE).getBoundingBox(chops);
+        VoxelShape voxelShape = Shapes.box(
+                box.minX,
+                box.minY,
+                box.minZ,
+                box.maxX,
+                box.maxY,
+                box.maxZ
         );
         return voxelShape;
     }
 
-    @SuppressWarnings({"deprecation", "NullableProblems"})
     @Override
-    public VoxelShape getRenderShape(BlockState state, IBlockReader world, BlockPos pos) {
-        return state.get(SHAPE).getOcclusionShape();
-    }
-
-    @Override
-    public boolean isSideInvisible(BlockState state, BlockState adjacentBlockState, Direction side) {
-        return true;
-    }
-
-    @Override
-    protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         builder.add(CHOPS, SHAPE);
     }
 
@@ -95,12 +89,12 @@ public class ChoppedLogBlock extends Block implements IChoppableBlock {
         if (numChops > getMaxNumChops()) {
             throw new IllegalArgumentException("Too many chops");
         }
-        return blockState.with(CHOPS, numChops);
+        return blockState.setValue(CHOPS, numChops);
     }
 
     @Override
     public int getNumChops(BlockState blockState) {
-        return blockState.get(CHOPS);
+        return blockState.getValue(CHOPS);
     }
 
     @Override
