@@ -1,54 +1,54 @@
-package ht.treechop.common.compat;
-
-import ht.treechop.TreeChopMod;
-import ht.treechop.common.config.ConfigHandler;
-import ht.treechop.common.event.ChopEvent;
-import ht.treechop.common.util.TickUtil;
-import net.minecraft.entity.Entity;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.entity.player.PlayerInteractEvent;
-import net.minecraftforge.eventbus.api.EventPriority;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@EventBusSubscriber(modid = TreeChopMod.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
-public class NoChopOnRightClick {
-
-    private static final Map<Entity, Long> lastRightClickTickByPlayers = new HashMap<>();
-    private static boolean enabled = false;
-
-    @SubscribeEvent
-    public static void commonSetup(FMLCommonSetupEvent event) {
-        if (ConfigHandler.COMMON.preventChoppingOnRightClick.get()) {
-            enable();
-        }
-    }
-
-    public static void enable() {
-        if (!enabled) {
-            MinecraftForge.EVENT_BUS.register(EventHandler.class);
-            enabled = true;
-        }
-    }
-
-    private static class EventHandler {
-        @SubscribeEvent(priority = EventPriority.HIGHEST)
-        public static void onBlockStartClick(PlayerInteractEvent.RightClickBlock event) {
-            long time = event.getWorld().getGameTime();
-            lastRightClickTickByPlayers.put(event.getPlayer(), time);
-        }
-
-        @SubscribeEvent
-        public static void onChop(ChopEvent.StartChopEvent event) {
-            long time = event.getWorld().getGameTime();
-            if (lastRightClickTickByPlayers.getOrDefault(event.getPlayer(), TickUtil.NEVER) == time) {
-                event.setCanceled(true);
-            }
-        }
-    }
-
-}
+//package ht.treechop.common.compat;
+//
+//import ht.treechop.TreeChopMod;
+//import ht.treechop.common.config.ConfigHandler;
+//import ht.treechop.api.ChopEvent;
+//import ht.treechop.common.util.TickUtil;
+//import net.minecraft.entity.Entity;
+//import net.minecraftforge.common.MinecraftForge;
+//import net.minecraftforge.event.entity.player.PlayerInteractEvent;
+//import net.minecraftforge.eventbus.api.EventPriority;
+//import net.minecraftforge.eventbus.api.SubscribeEvent;
+//import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+//import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+//
+//import java.util.HashMap;
+//import java.util.Map;
+//
+//@EventBusSubscriber(modid = TreeChopMod.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
+//public class NoChopOnRightClick {
+//
+//    private static final Map<Entity, Long> lastRightClickTickByPlayers = new HashMap<>();
+//    private static boolean enabled = false;
+//
+//    @SubscribeEvent
+//    public static void commonSetup(FMLCommonSetupEvent event) {
+//        if (ConfigHandler.COMMON.preventChoppingOnRightClick.get()) {
+//            enable();
+//        }
+//    }
+//
+//    public static void enable() {
+//        if (!enabled) {
+//            MinecraftForge.EVENT_BUS.register(EventHandler.class);
+//            enabled = true;
+//        }
+//    }
+//
+//    private static class EventHandler {
+//        @SubscribeEvent(priority = EventPriority.HIGHEST)
+//        public static void onBlockStartClick(PlayerInteractEvent.RightClickBlock event) {
+//            long time = event.getWorld().getGameTime();
+//            lastRightClickTickByPlayers.put(event.getPlayer(), time);
+//        }
+//
+//        @SubscribeEvent
+//        public static void onChop(ChopEvent.StartChopEvent event) {
+//            long time = event.getWorld().getGameTime();
+//            if (lastRightClickTickByPlayers.getOrDefault(event.getPlayer(), TickUtil.NEVER) == time) {
+//                event.setCanceled(true);
+//            }
+//        }
+//    }
+//
+//}
