@@ -1,12 +1,12 @@
 package ht.treechop.common.config;
 
-import net.minecraft.util.IStringSerializable;
+import net.minecraft.util.StringRepresentable;
 
 import java.util.function.Function;
 
 import static java.lang.Math.log;
 
-public enum ChopCountingAlgorithm implements IStringSerializable {
+public enum ChopCountingAlgorithm implements StringRepresentable {
     LINEAR(
             numBlocks -> {
                 double x = (double) numBlocks;
@@ -42,7 +42,7 @@ public enum ChopCountingAlgorithm implements IStringSerializable {
     }
 
     @Override
-    public String getString() {
+    public String getSerializedName() {
         return name();
     }
 }
