@@ -1,7 +1,7 @@
 package ht.treechop.common.config.item;
 
-import net.minecraft.item.Item;
-import net.minecraft.tags.ITagCollection;
+import net.minecraft.tags.TagCollection;
+import net.minecraft.world.item.Item;
 import net.minecraftforge.registries.IForgeRegistry;
 
 import java.util.Collections;
@@ -15,7 +15,7 @@ public class MalformedItemIdentifier extends ItemIdentifier {
     }
 
     @Override
-    public List<Item> resolve(ITagCollection<Item> tags, IForgeRegistry<Item> registry) {
+    public List<Item> resolve(TagCollection<Item> tags, IForgeRegistry<Item> registry) {
         return Collections.emptyList();
     }
 
