@@ -1,19 +1,19 @@
 package ht.treechop.common.util;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
 
 public class TreeBlock extends WorldBlock {
     private final boolean chopped;
 
-    public TreeBlock(World world, BlockPos pos, BlockState state, boolean chopped) {
-        super(world, pos, state);
+    public TreeBlock(Level level, BlockPos pos, BlockState state, boolean chopped) {
+        super(level, pos, state);
         this.chopped = true;
     }
 
-    public TreeBlock(World world, BlockPos pos, BlockState state) {
-        this(world, pos, state, false);
+    public TreeBlock(Level level, BlockPos pos, BlockState state) {
+        this(level, pos, state, false);
     }
 
     public boolean wasChopped() {
