@@ -1,10 +1,10 @@
 package ht.treechop.common.util;
 
-import net.minecraft.dispenser.IPosition;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.math.vector.Vector3f;
+import com.mojang.math.Vector3d;
+import com.mojang.math.Vector3f;
+import net.minecraft.core.Position;
 
-public class Vector3 implements IPosition {
+public class Vector3 implements Position {
 
     public static final Vector3 ZERO = new Vector3(0, 0, 0);
 
@@ -19,27 +19,27 @@ public class Vector3 implements IPosition {
     }
 
     public Vector3(Vector3f vector3f) {
-        this(vector3f.getX(), vector3f.getY(), vector3f.getZ());
+        this(vector3f.x(), vector3f.y(), vector3f.z());
     }
 
-    public Vector3(IPosition position) {
-        this(position.getX(), position.getY(), position.getZ());
+    public Vector3(Position position) {
+        this(position.x(), position.y(), position.z());
     }
 
-    public double getX() {
+    public double x() {
         return x;
     }
 
-    public double getY() {
+    public double y() {
         return y;
     }
 
-    public double getZ() {
+    public double z() {
         return z;
     }
 
     public Vector3 withX(double x) {
-        return new Vector3(x, this.getY(), this.getZ());
+        return new Vector3(x, this.y(), this.z());
     }
 
     public Vector3 withX(Number x) {
@@ -47,7 +47,7 @@ public class Vector3 implements IPosition {
     }
 
     public Vector3 withY(double y) {
-        return new Vector3(this.getX(), y, this.getZ());
+        return new Vector3(this.x(), y, this.z());
     }
 
     public Vector3 withY(Number y) {
@@ -55,7 +55,7 @@ public class Vector3 implements IPosition {
     }
 
     public Vector3 withZ(double z) {
-        return new Vector3(this.getX(), this.getY(), z);
+        return new Vector3(this.x(), this.y(), z);
     }
 
     public Vector3 withZ(Number z) {
@@ -63,7 +63,7 @@ public class Vector3 implements IPosition {
     }
 
     public Vector3 scale(double factor) {
-        return new Vector3(this.getX() * factor, this.getY() * factor, this.getZ() * factor);
+        return new Vector3(this.x() * factor, this.y() * factor, this.z() * factor);
     }
 
     public Vector3 scale(Number factor) {
@@ -72,17 +72,17 @@ public class Vector3 implements IPosition {
 
     public Vector3 add(Vector3 vector3) {
         return new Vector3(
-                this.getX() + vector3.getX(),
-                this.getY() + vector3.getY(),
-                this.getZ() + vector3.getZ()
+                this.x() + vector3.x(),
+                this.y() + vector3.y(),
+                this.z() + vector3.z()
         );
     }
 
     public Vector3 add(double amount) {
         return new Vector3(
-                this.getX() + amount,
-                this.getY() + amount,
-                this.getZ() + amount
+                this.x() + amount,
+                this.y() + amount,
+                this.z() + amount
         );
     }
 
@@ -91,11 +91,11 @@ public class Vector3 implements IPosition {
     }
 
     public Vector3f asVector3f() {
-        return new Vector3f((float)this.getX(), (float)this.getY(), (float)this.getZ());
+        return new Vector3f((float)this.x(), (float)this.y(), (float)this.z());
     }
 
     public Vector3d asVector3d() {
-        return new Vector3d(this.getX(), this.getY(), this.getZ());
+        return new Vector3d(this.x(), this.y(), this.z());
     }
 
 }
