@@ -1,29 +1,29 @@
 package ht.treechop.common.util;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class WorldBlock {
-    private final World world;
+    private final Level level;
     private final BlockPos pos;
     private final BlockState state;
-    private final TileEntity tileEntity;
+    private final BlockEntity tileEntity;
 
-    public WorldBlock(World world, BlockPos pos, BlockState state, TileEntity tileEntity) {
-        this.world = world;
+    public WorldBlock(Level level, BlockPos pos, BlockState state, BlockEntity tileEntity) {
+        this.level = level;
         this.pos = pos;
         this.state = state;
         this.tileEntity = tileEntity;
     }
 
-    public WorldBlock(World world, BlockPos pos, BlockState state) {
-        this(world, pos, state, null);
+    public WorldBlock(Level level, BlockPos pos, BlockState state) {
+        this(level, pos, state, null);
     }
 
     public BlockPos getPos() {
@@ -34,21 +34,21 @@ public class WorldBlock {
         return state;
     }
 
-    public TileEntity getTileEntity() {
+    public BlockEntity getTileEntity() {
         return tileEntity;
     }
 
-    public World getWorld() {
-        return world;
+    public Level getWorld() {
+        return level;
     }
 
-    public static WorldBlock fromWorld(World world, BlockPos pos) {
-        return new WorldBlock(world, pos, world.getBlockState(pos), world.getTileEntity(pos));
+    public static WorldBlock fromWorld(Level level, BlockPos pos) {
+        return new WorldBlock(level, pos, level.getBlockState(pos), level.getBlockEntity(pos));
     }
 
-    public static List<WorldBlock> fromWorld(World world, Collection<BlockPos> positions) {
+    public static List<WorldBlock> fromWorld(Level level, Collection<BlockPos> positions) {
         return positions.stream()
-                .map(pos -> fromWorld(world, pos))
+                .map(pos -> fromWorld(level, pos))
                 .collect(Collectors.toList());
     }
 }
